Move FisheyeProjection to scene/graphics for shared use
